Split snapToGrid extension
